+++ src/main/java/blusunrize/immersiveengineering/common/gui/IESlot.java
@@ -9,11 +9,13 @@
 package blusunrize.immersiveengineering.common.gui;
 
 import blusunrize.immersiveengineering.ImmersiveEngineering;
-import blusunrize.immersiveengineering.api.crafting.*;
+import blusunrize.immersiveengineering.api.crafting.ArcFurnaceRecipe;
+import blusunrize.immersiveengineering.api.crafting.BlastFurnaceFuel;
+import blusunrize.immersiveengineering.api.crafting.BlueprintCraftingRecipe;
+import blusunrize.immersiveengineering.api.crafting.ClocheRecipe;
 import blusunrize.immersiveengineering.api.shader.CapabilityShader;
 import blusunrize.immersiveengineering.api.shader.IShaderItem;
 import blusunrize.immersiveengineering.api.tool.IConfigurableTool;
-import blusunrize.immersiveengineering.api.tool.IDrillHead;
 import blusunrize.immersiveengineering.api.tool.IUpgrade;
 import blusunrize.immersiveengineering.api.tool.IUpgradeableTool;
 import blusunrize.immersiveengineering.common.items.BulletItem;
@@ -173,6 +175,7 @@ public abstract class IESlot extends Slot
 	public static class WithPredicate extends SlotItemHandler
 	{
 		final Predicate<ItemStack> predicate;
+
 		public WithPredicate(IItemHandler inv, int id, int x, int y, Predicate<ItemStack> predicate)
 		{
 			super(inv, id, x, y);
@@ -199,9 +202,10 @@ public abstract class IESlot extends Slot
 		final boolean preventDoubles;
 		final Container container;
 		final Supplier<World> getWorld;
+		final Supplier<PlayerEntity> getPlayer;
 
 		public Upgrades(Container container, IItemHandler inv, int id, int x, int y, String type, ItemStack upgradeableTool,
-						boolean preventDoubles, Supplier<World> getWorld)
+						boolean preventDoubles, Supplier<World> getWorld, Supplier<PlayerEntity> getPlayer)
 		{
 			super(inv, id, x, y);
 			this.container = container;
@@ -209,6 +213,7 @@ public abstract class IESlot extends Slot
 			this.upgradeableTool = upgradeableTool;
 			this.preventDoubles = preventDoubles;
 			this.getWorld = getWorld;
+			this.getPlayer = getPlayer;
 		}
 
 		@Override
@@ -228,9 +233,22 @@ public abstract class IESlot extends Slot
 		}
 
 		@Override
+		public ItemStack onTake(PlayerEntity thePlayer, ItemStack stack)
+		{
+			stack = ((IUpgradeableTool)upgradeableTool.getItem()).removeUpgrade(upgradeableTool, thePlayer, stack);
+			stack = super.onTake(thePlayer, stack);
+			return stack;
+		}
+
+		@Override
 		public void onSlotChanged()
 		{
-			((IUpgradeableTool)upgradeableTool.getItem()).recalculateUpgrades(upgradeableTool, getWorld.get());
+			super.onSlotChanged();
+			((IUpgradeableTool)upgradeableTool.getItem()).recalculateUpgrades(upgradeableTool, getWorld.get(), getPlayer.get());
+			if(container instanceof ModWorkbenchContainer)
+				((ModWorkbenchContainer)container).rebindSlots();
+			else if(container instanceof MaintenanceKitContainer)
+				((MaintenanceKitContainer)container).updateSlots();
 		}
 	}