+++ src/main/java/blusunrize/immersiveengineering/common/items/BuzzsawItem.java
@@ -25,6 +25,8 @@ import blusunrize.immersiveengineering.client.render.IEOBJItemRenderer;
 import blusunrize.immersiveengineering.common.gui.IESlot;
 import blusunrize.immersiveengineering.common.items.IEItemInterfaces.IAdvancedFluidItem;
 import blusunrize.immersiveengineering.common.items.IEItemInterfaces.IScrollwheel;
+import blusunrize.immersiveengineering.common.items.IEItems.Misc;
+import blusunrize.immersiveengineering.common.items.ToolUpgradeItem.ToolUpgrade;
 import blusunrize.immersiveengineering.common.util.ItemNBTHelper;
 import blusunrize.immersiveengineering.common.util.Utils;
 import blusunrize.immersiveengineering.common.util.fluids.IEItemFluidHandler;
@@ -178,7 +180,7 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 	}
 
 	@Override
-	public Slot[] getWorkbenchSlots(Container container, ItemStack stack, Supplier<World> getWorld)
+	public Slot[] getWorkbenchSlots(Container container, ItemStack stack, Supplier<World> getWorld, Supplier<PlayerEntity> getPlayer)
 	{
 		IItemHandler inv = stack.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY, null)
 				.orElseThrow(RuntimeException::new);
@@ -186,8 +188,8 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 			return new Slot[]
 					{
 							new IESlot.WithPredicate(inv, 0, 88, 22, (itemStack) -> sawblades.contains(itemStack.getItem())),
-							new IESlot.Upgrades(container, inv, 1, 88, 52, "BUZZSAW", stack, true, getWorld),
-							new IESlot.Upgrades(container, inv, 2, 108, 52, "BUZZSAW", stack, true, getWorld),
+							new IESlot.Upgrades(container, inv, 1, 88, 52, "BUZZSAW", stack, true, getWorld, getPlayer),
+							new IESlot.Upgrades(container, inv, 2, 108, 52, "BUZZSAW", stack, true, getWorld, getPlayer),
 							new IESlot.WithPredicate(inv, 3, 108, 12, (itemStack) -> sawblades.contains(itemStack.getItem())),
 							new IESlot.WithPredicate(inv, 4, 108, 32, (itemStack) -> sawblades.contains(itemStack.getItem()))
 					};
@@ -195,8 +197,8 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 			return new Slot[]
 					{
 							new IESlot.WithPredicate(inv, 0, 98, 22, (itemStack) -> sawblades.contains(itemStack.getItem())),
-							new IESlot.Upgrades(container, inv, 1, 88, 52, "BUZZSAW", stack, true, getWorld),
-							new IESlot.Upgrades(container, inv, 2, 108, 52, "BUZZSAW", stack, true, getWorld)
+							new IESlot.Upgrades(container, inv, 1, 88, 52, "BUZZSAW", stack, true, getWorld, getPlayer),
+							new IESlot.Upgrades(container, inv, 2, 108, 52, "BUZZSAW", stack, true, getWorld, getPlayer)
 					};
 	}
 
@@ -204,6 +206,22 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 	public boolean canModify(ItemStack stack)
 	{
 		return true;
+	}
+
+	@Override
+	public ItemStack removeUpgrade(ItemStack stack, PlayerEntity player, ItemStack upgrade)
+	{
+		if(upgrade.getItem()==Misc.toolUpgrades.get(ToolUpgrade.BUZZSAW_SPAREBLADES))
+			for(int i = 1; i <= 2; i++)
+			{
+				ItemStack sawblade = getSawblade(stack, i);
+				if(!sawblade.isEmpty())
+				{
+					ItemNBTHelper.setItemStack(upgrade, "sawblade"+i, sawblade);
+					setSawblade(stack, ItemStack.EMPTY, i);
+				}
+			}
+		return upgrade;
 	}
 
 	@Override
@@ -217,9 +235,24 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 	}
 
 	@Override
-	public void recalculateUpgrades(ItemStack stack, World w)
+	public void recalculateUpgrades(ItemStack stack, World w, PlayerEntity player)
 	{
-		super.recalculateUpgrades(stack, w);
+		super.recalculateUpgrades(stack, w, player);
+		LazyOptional<IItemHandler> invCap = stack.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY, null);
+		invCap.ifPresent(inv -> {
+			for(int iUpgrade = 1; iUpgrade <= 2; iUpgrade++)
+			{
+				ItemStack upgrade = inv.getStackInSlot(iUpgrade);
+				if(upgrade.getItem()==Misc.toolUpgrades.get(ToolUpgrade.BUZZSAW_SPAREBLADES))
+					for(int i = 1; i <= 2; i++)
+						if(ItemNBTHelper.hasKey(upgrade, "sawblade"+i))
+						{
+							ItemStack sawblade = ItemNBTHelper.getItemStack(upgrade, "sawblade"+i);
+							setSawblade(stack, sawblade, i);
+							ItemNBTHelper.remove(upgrade, "sawblade"+i);
+						}
+			}
+		});
 		FluidStack fs = getFluid(stack);
 		if(fs!=null&&fs.getAmount() > this.getCapacity(stack, 2000))
 		{
@@ -401,7 +434,7 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 
 	private void consumeDurability(ItemStack stack, World world, BlockState state, BlockPos pos, LivingEntity living)
 	{
-		if(state==null || state.getBlockHardness(world, pos)!=0.0f)
+		if(state==null||state.getBlockHardness(world, pos)!=0.0f)
 		{
 			int dmg = state==null||ForgeHooks.isToolEffective(world, pos, stack)||isEffective(stack, state.getMaterial())?1: 3;
 			ItemStack sawblade = getSawblade(stack);
@@ -592,7 +625,7 @@ public class BuzzsawItem extends UpgradeableToolItem implements IAdvancedFluidIt
 			@SubscribeEvent
 			public void onTick(TickEvent.WorldTickEvent event)
 			{
-				if(event.phase==Phase.START && event.world == world)
+				if(event.phase==Phase.START&&event.world==world)
 				{
 					breakFromList(closedList, 5, world, player, stack);
 					if(closedList.isEmpty())